feat: default pkg handler
